% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcZ}
\alias{calcZ}
\title{\code{calcZ} calculates the Z-score for each element based
on input mean and standard deviation matrices}
\usage{
calcZ(meanMat, sdMat)
}
\arguments{
\item{meanMat}{matrix of mean values}

\item{sdMat}{matrix of standard deviation values}
}
\description{
\code{calcZ} calculates the Z-score for each element based
on input mean and standard deviation matrices
}

