% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GAPS}
\alias{GAPS}
\title{\code{GAPS} is retired as of version 2, use gapsRun instead}
\usage{
GAPS(data, unc, outputDir, outputBase = "", sep = "\\t",
  isPercentError = FALSE, numPatterns, MaxAtomsA = 2^32, alphaA = 0.01,
  MaxAtomsP = 2^32, alphaP = 0.01, SAIter = 1e+09, iter = 5e+08,
  thin = -1, verbose = TRUE, keepChain = FALSE)
}
\description{
\code{GAPS} is retired as of version 2, use gapsRun instead
}

