% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSmoothPatterns}
\alias{plotSmoothPatterns}
\title{\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively}
\usage{
plotSmoothPatterns(P, x = NULL, breaks = NULL, breakStyle = T,
  orderP = !all(is.null(x)), plotPTS = F, pointCol = "black",
  lineCol = "grey", add = F, ...)
}
\arguments{
\item{P}{the mean A matrix}

\item{x}{optional variables}

\item{breaks}{breaks in plots}

\item{breakStyle}{style of breaks}

\item{orderP}{whether to order patterns}

\item{plotPTS}{whether to plot points on lines}

\item{pointCol}{color of points}

\item{lineCol}{color of line}

\item{add}{add}{logical specifying if bars should be added to an already existing plot; defaults to \code{FALSE}}

\item{...}{arguments to be passed to/from other methods.  For the default method these can include further arguments (such as \code{axes}, \code{asp} and \code{main}) and graphical parameters (see \code{par}) which are passed to \code{plot.window()}, \code{title()} and \code{axis}}

}
\description{
\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively
}

