% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSmoothPatterns}
\alias{plotSmoothPatterns}
\title{\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively}
\usage{
plotSmoothPatterns(P, x = NULL, breaks = NULL, breakStyle = T,
  orderP = !all(is.null(x)), plotPTS = F, pointCol = "black",
  lineCol = "grey", add = F, ...)
}
\arguments{
\item{P}{the mean A matrix}

\item{x}{optional variables}

\item{breaks}{breaks in plots}

\item{breakStyle}{style of breaks}

\item{orderP}{whether to order patterns}

\item{plotPTS}{whether to plot points on lines}

\item{pointCol}{color of points}

\item{lineCol}{color of line}
}
\description{
\code{plotSmoothPatterns} plots the output A and P matrices as a
heatmap and line plot respectively
}

